requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Cors==4.0.0
pymongo==4.6.3
selenium==4.19.0
python-dotenv==1.0.1
setuptools==69.5.1
typing_extensions==4.11.0
# gunicorn==20.1.0
# Flask
Gunicorn
webdriver-manager==4.0.1